chore(examples): refactor
