+++ package.json
@@ -4,9 +4,7 @@
   "description": "Example codes for creating your own services with Levain GraphQL APIs.",
   "main": "src/withdrawal.ts",
   "scripts": {
-    "start:deposit": "nodemon src/generate-deposit-addresses.ts",
-    "start:withdrawal": "nodemon src/api-cosigning.ts",
-    "start:wallet": "nodemon src/create-wallet.ts",
+    "start": "nodemon src/app.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
