+++ src/app.ts
@@ -1,22 +1,29 @@
 import express from "express";
 import dotenv from "dotenv";
+import apiCosigningRoute from "./api/api-cosigning";
+import walletRoute from "./api/create-wallet";
+import depositAddressRoute from "./api/generate-deposit-addresses";
 
 dotenv.config();
 
 const app = express();
 const port = 3000;
 
-// Endpoint to process withdrawal
 app.get("/", async (req, res) => {
   try {
     res.status(200).json({
-      message: "Demo server running OK",
+      message: "Demo service running OK.",
     });
   } catch (error) {
     console.log(error);
     res.status(500).json({ error: error });
   }
 });
+
+app.use(apiCosigningRoute);
+app.use(walletRoute);
+app.use(depositAddressRoute);
+
 
 app.listen(port, () => {
   console.log(
